refactor out project, filter, ordering generator from SparkPlan
