+++ sql/core/src/main/scala/org/apache/spark/sql/execution/aggregate/SortBasedAggregate.scala
@@ -87,8 +87,7 @@ case class SortBasedAggregate(
           completeAggregateAttributes,
           initialInputBufferOffset,
           resultExpressions,
-          newMutableProjection _,
-          newProjection _,
+          codegenEnabled,
           child.output,
           iter,
           outputsUnsafeRows,
